by golly i think i did it
