+++ generators/build.gradle.kts
@@ -1,7 +1,8 @@
 plugins {
-  kotlin("jvm")
+  kotlin("multiplatform")
   kotlin("plugin.serialization") version "1.6.10"
-  id("io.bkbn.sourdough.library.jvm") version "0.6.0"
+  id("io.kotest.multiplatform") version "5.1.0"
+  id("io.bkbn.sourdough.library.mpp") version "0.6.0"
   id("io.gitlab.arturbosch.detekt") version "1.19.0"
   id("com.adarshr.test-logger") version "3.1.0"
   id("org.jetbrains.dokka")
@@ -27,22 +28,42 @@ dependencies {
   implementation(group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-json", version = "1.3.2")
   detektPlugins(group = "io.gitlab.arturbosch.detekt", name = "detekt-formatting", version = "1.19.0")
 }
-repositories {
-  mavenCentral()
-}
 
-testing {
-  suites {
-    named("test", JvmTestSuite::class) {
-      useJUnitJupiter()
+kotlin {
+  sourceSets {
+    val commonMain by getting {
+      resources.srcDirs("resources")
       dependencies {
-        // Satisfaketion
-        implementation(projects.satisfaketionMutators)
-
-        // Kotest
-        implementation("io.kotest:kotest-runner-junit5-jvm:5.1.0")
-        implementation("io.kotest:kotest-assertions-core-jvm:5.1.0")
+        implementation(kotlin("stdlib"))
+        implementation(projects.satisfaketionCore)
+        implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.3.2")
+        implementation("com.squareup.okio:okio:3.0.0")
+        implementation("co.touchlab:kermit:1.0.3")
       }
+    }
+    val commonTest by getting {
+      dependencies {
+        implementation("io.kotest:kotest-assertions-core:5.0.3")
+        implementation("io.kotest:kotest-framework-engine:5.0.3")
+      }
+    }
+    val jvmMain by getting {
+      resources.srcDirs("resources")
+      dependencies {
+        implementation("io.kotest:kotest-runner-junit5-jvm:5.0.3")
+      }
+    }
+    val jvmTest by getting
+    val jsMain by getting {
+      resources.srcDirs("resources")
+      dependencies {
+        implementation("com.squareup.okio:okio-nodefilesystem:3.0.0")
+      }
+    }
+    val jsTest by getting
+    val nativeMain by getting
+    val nativeTest by getting {
+      resources.srcDirs("resources")
     }
   }
 }
