+++ generators/src/commonMain/kotlin/io/github/unredundant/satisfaketion/generators/common/Utils.kt
@@ -1,0 +1,7 @@
+package io.github.unredundant.satisfaketion.generators.common
+
+import okio.Path
+
+expect object Reader {
+  fun read(path: Path): String
+}
