+++ generators/src/commonTest/kotlin/io/github/unredundant/satisfaketion/generators/Placeholder.kt
@@ -1,0 +1,4 @@
+package io.github.unredundant.satisfaketion.generators
+
+object Placeholder {
+}
