+++ generators/src/commonTest/kotlin/io/github/unredundant/satisfaketion/generators/en/EnglishNameTest.kt
@@ -1,0 +1,6 @@
+package io.github.unredundant.satisfaketion.generators.en
+
+import io.kotest.core.spec.style.DescribeSpec
+
+class EnglishNameTest : DescribeSpec({
+})
