--- generators/src/main/kotlin/io/github/unredundant/satisfaketion/generators/common/Utils.kt
@@ -1,7 +1,0 @@
-package io.github.unredundant.satisfaketion.generators.common
-
-object Utils {
-  fun getFile(fileName: String): String {
-    return this::class.java.classLoader.getResource(fileName).readText()
-  }
-}
