+++ generators/src/resources/barcode.json
