+++ generators/src/resources/beer.json
